+++ instrumentation/opentelemetry-instrumentation-asgi/tests/test_asgi_middleware.py
@@ -164,7 +164,7 @@ class TestAsgiApplication(AsgiTestBase):
         outputs = self.get_all_output()
         self.validate_outputs(outputs)
 
-    def test_wsgi_not_recording(self):
+    def test_asgi_not_recording(self):
         mock_tracer = mock.Mock()
         mock_span = mock.Mock()
         mock_span.is_recording.return_value = False
@@ -312,8 +312,12 @@ class TestAsgiAttributes(unittest.TestCase):
 
     def test_request_attributes(self):
         self.scope["query_string"] = b"foo=bar"
+        headers = []
+        headers.append(("host".encode("utf8"), "test".encode("utf8")))
+        self.scope["headers"] = headers
 
         attrs = otel_asgi.collect_request_attributes(self.scope)
+
         self.assertDictEqual(
             attrs,
             {
@@ -324,6 +328,7 @@ class TestAsgiAttributes(unittest.TestCase):
                 "http.url": "http://127.0.0.1/?foo=bar",
                 "host.port": 80,
                 "http.scheme": "http",
+                "http.server_name": "test",
                 "http.flavor": "1.0",
                 "net.peer.ip": "127.0.0.1",
                 "net.peer.port": 32767,