+++ instrumentation/opentelemetry-instrumentation-wsgi/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Return `None` for `CarrierGetter` if key not found
+  ([#1374](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/233))
+
 ## Version 0.13b0
 
 Released 2020-09-17